extracts onDrop into a util function
