+++ src/component/List/index.js
@@ -2,7 +2,13 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { v4 as uuid } from 'uuid';
 import Card from '../Card';
-import { removeItemById, addCardToState, deleteCardFromState, updateCardFromState } from '../../utils/index';
+import {
+  removeItemById,
+  addCardToState,
+  deleteCardFromState,
+  updateCardFromState,
+  updateLists,
+} from '../../utils/index';
 
 export default function List({ list, setBoardState }) {
   const [addCardInput, setAddCardInput] = useState('');
@@ -22,31 +28,7 @@ export default function List({ list, setBoardState }) {
   function onDrop(e, targetList) {
     e.preventDefault();
     const cardInfo = e.dataTransfer.getData('draggedCardData');
-
-    const { cardId, listId } = JSON.parse(cardInfo);
-
-    const updateState = state => {
-      const originList = state.find(listValue => listValue.id === listId);
-      const destinationList = state.find(listValue => listValue.id === targetList);
-
-      const updatedOriginList = {
-        ...originList,
-        cards: removeItemById(originList.cards, cardId),
-      };
-      const updatedDestinationList = {
-        ...destinationList,
-        cards: [...destinationList.cards, originList.cards.find(item => item.id === cardId)],
-      };
-
-      return state.map(listValue => {
-        if (originList.id === destinationList.id) return listValue; // Dragged to the same list, don't update state.
-        if (listValue.id === originList.id) return updatedOriginList;
-        if (listValue.id === destinationList.id) return updatedDestinationList;
-        return listValue;
-      });
-    };
-
-    setBoardState(updateState);
+    setBoardState(state => updateLists(state, JSON.parse(cardInfo), targetList));
   }
 
   const onAddNewCard = e => {
